video-generator/requirements.txt
pillow
gtts
moviepy
ffmpeg-python
pyttsx3
flask
streamlit
